#2721: moved to runtime
