+++ eo-runtime/src/test/resources/org.eolang/snippets/all.yaml
