+++ eo-runtime/src/test/resources/org.eolang/snippets/fibo.yaml
