+++ eo-runtime/src/test/resources/org.eolang/snippets/ifthenelse.yaml
