+++ eo-runtime/src/test/resources/org.eolang/snippets/parenting.yaml
