+++ eo-runtime/src/test/resources/org.eolang/snippets/simple.yaml
