+++ pom.xml
@@ -211,6 +211,12 @@ SOFTWARE.
         <scope>test</scope>
       </dependency>
       <dependency>
+        <groupId>com.yegor256</groupId>
+        <artifactId>jping</artifactId>
+        <version>0.0.1</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
         <groupId>org.slf4j</groupId>
         <artifactId>slf4j-reload4j</artifactId>
         <version>2.0.9</version>
